Fix error of TARGET_OTA_ASSERT_DEVICE
In the Makefile, any line following the target whose first character is
a tab is recognized as a command, unless it is a continuation of a
previous line using a backslash.

Thus, since the $(hide) is currently preceded by a tab, it is considered
to be a continuation of the previous line, as if it were a
ifneq ($(TARGET_OTA_ASSERT_DEVICE),) $(hide) echo
"ota_override_device=$(TARGET_OTA_ASSERT_DEVICE)" >> $@
is written on a single line.

Therefore, the following line cannot be executed correctly
$(hide) echo "ota_override_device=$(TARGET_OTA_ASSERT_DEVICE)" >> $@

Signed-off-by: BrightTwikling <DF.SJbbba5@gmail.com>